+++ src/ONeWD.h
@@ -71,6 +71,12 @@ protected:
 
     double          CalculateLuminosityOnPhase() const                              { return CalculateLuminosityOnPhase(m_Mass, m_Age, m_Metallicity); }    // Use class member variables
 
+    DBL_DBL         CalculateMassAcceptanceRate(const double p_DonorMassRate,
+                                                const bool   p_IsHeRich);  
+    DBL_DBL         CalculateMassAcceptanceRate(const double p_DonorMassRate,
+                                                const double p_AccretorMassRate,
+                                                const bool   p_IsHeRich)            { return CalculateMassAcceptanceRate(p_DonorMassRate, p_IsHeRich); }        // Ignore the input accretion rate for WDs
+    
     STELLAR_TYPE    EvolveToNextPhase();
     bool            IsSupernova() const;                                             
     bool            ShouldEvolveOnPhase() const;